Задача 7 выполнена.  Реализовать класс BeanUtils по документации
